requirements.txt
rich
bcrypt
sqlalchemy
pymysql
cryptography
pytest
pytest-cov
python-dotenv
flask
flask-sqlalchemy
